Apply changes from PR comments. Godoc for tests
